requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
databases==0.8.0
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a35
starlette==0.27.0
tomli==2.0.1
types-passlib==1.7.7.13
types-pyasn1==0.4.0.6
types-python-jose==3.3.4.8
typing_extensions==4.8.0
uvicorn==0.23.2
